requirements.txt
appdirs==1.4.3
asgiref==3.2.3
astroid==2.3.3
atomicwrites==1.3.0
attrs==19.3.0
autopep8==1.5.1
certifi==2019.11.28
chardet==3.0.4
colorama==0.4.1
distlib==0.3.0
Django==3.0
django-phone-field==1.8.0
filelock==3.0.12
get==2019.4.13
idna==2.9
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
more-itertools==8.2.0
packaging==20.1
Pillow==6.2.1
pluggy==0.13.1
post==2019.4.13
public==2019.4.13
py==1.8.1
pycodestyle==2.5.0
pylint==2.4.4
pyparsing==2.4.6
pytest==5.3.5
pytz==2019.3
query-string==2019.4.13
request==2019.4.13
requests==2.23.0
six==1.13.0
sqlparse==0.3.0
urllib3==1.25.8
virtualenv==20.0.3
virtualenvwrapper-win==1.2.6
wcwidth==0.1.8
wrapt==1.11.2
